pyproject.toml
[project]
name = "summarization-pipeline"
version = "0.1.0"
description = "Streamlit web app for summarizing news articles using bart-large-cnn"
requires-python = ">=3.12"
dependencies = [
    "lxml_html_clean",
    "newspaper4k",
    "nltk",
    "streamlit",
    "torch",
    "transformers",
]

[dependency-groups]
dev = [
    "ty",
    "pytest",
    "ruff",
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ty.environment]
python-version = "3.12"
